requirements.txt
ultralytics
customtkinter
opencv-python
pillow
numpy